requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cairocffi==0.9.0
CairoSVG==1.0.22
cffi==1.1.0
cssselect2==0.2.1
decorator==4.4.0
Django==1.11.20
django-filter==2.0.0
djangorestframework==3.9.2
enum34==1.1.6
html5lib==1.0.1
ipdb==0.11
ipython==5.8.0
ipython-genutils==0.2.0
pathlib==1.0.1
pathlib2==2.3.3
pdfrw==0.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==2.2.2
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
Pyphen==0.9.5
pytz==2018.9
qrcode==6.1
scandir==1.10.0
simplegeneric==0.8.1
six==1.12.0
tinycss2==0.6.1
traitlets==4.3.2
virtualenv==16.4.3
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5.1